subdomain_brute_forcer/requirements.txt
dnspython
termcolor
argparse
google
googlesearch-python

